+++ src/main/java/Calculator/domain/Operator.java
@@ -1,7 +1,7 @@
 /*
 * 연산자를 반환하는 클래스
 * */
-package Calculator;
+package Calculator.domain;
 
 import java.util.function.BinaryOperator;
