+++ src/test/java/Calculator/domain/DelimiterTest.java
@@ -1,9 +1,10 @@
-package Calculator;
+package Calculator.domain;
 
+import Calculator.domain.Delimiter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static Calculator.Delimiter.DEFAULT_DELIMITER;
+import static Calculator.domain.Delimiter.DEFAULT_DELIMITER;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class DelimiterTest {
