+++ src/test/java/Calculator/domain/OperandsTest.java
@@ -1,5 +1,8 @@
-package Calculator;
+package Calculator.domain;
 
+import Calculator.domain.Operand;
+import Calculator.domain.Operands;
+import Calculator.domain.Operator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
