+++ src/test/java/Calculator/util/StringUtilsTest.java
@@ -1,5 +1,6 @@
-package Calculator;
+package Calculator.util;
 
+import Calculator.domain.Delimiter;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
